// MCP API/gen/cmp/types/v4/evm_address_pb.ts
// @generated by protoc-gen-es v2.10.1 with parameter "target=ts,import_extension=.js"
// @generated from file cmp/types/v4/evm_address.proto (package cmp.types.v4, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_buf_validate_validate } from "../../../buf/validate/validate_pb.js";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file cmp/types/v4/evm_address.proto.
 */
export const file_cmp_types_v4_evm_address: GenFile = /*@__PURE__*/
  fileDesc("Ch5jbXAvdHlwZXMvdjQvZXZtX2FkZHJlc3MucHJvdG8SDGNtcC50eXBlcy52NCI5CgpFVk1BZGRyZXNzEisKB2FkZHJlc3MYASABKAlCGrpIF3IVMhNeMHhbMC05YS1mQS1GXXs0MH0kYgZwcm90bzM", [file_buf_validate_validate]);

/**
 * Address type that represents an EVM address in hex format
 *
 * Example: "0x1234567890abcdef1234567890abcdef12345678"
 *
 * ![Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/evm_address.proto.dot.xs.svg)
 *
 * [Open Message Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/evm_address.proto.dot.svg)
 *
 * @generated from message cmp.types.v4.EVMAddress
 */
export type EVMAddress = Message<"cmp.types.v4.EVMAddress"> & {
  /**
   * @generated from field: string address = 1;
   */
  address: string;
};

/**
 * Describes the message cmp.types.v4.EVMAddress.
 * Use `create(EVMAddressSchema)` to create a new message.
 */
export const EVMAddressSchema: GenMessage<EVMAddress> = /*@__PURE__*/
  messageDesc(file_cmp_types_v4_evm_address, 0);