// MCP API/gen/cmp/types/v4/contact_info_pb.ts
// @generated by protoc-gen-es v2.10.1 with parameter "target=ts,import_extension=.js"
// @generated from file cmp/types/v4/contact_info.proto (package cmp.types.v4, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_buf_validate_validate } from "../../../buf/validate/validate_pb.js";
import type { Address } from "./address_pb.js";
import { file_cmp_types_v4_address } from "./address_pb.js";
import type { Email } from "./email_pb.js";
import { file_cmp_types_v4_email } from "./email_pb.js";
import type { Link } from "./link_pb.js";
import { file_cmp_types_v4_link } from "./link_pb.js";
import type { Phone } from "./phone_pb.js";
import { file_cmp_types_v4_phone } from "./phone_pb.js";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file cmp/types/v4/contact_info.proto.
 */
export const file_cmp_types_v4_contact_info: GenFile = /*@__PURE__*/
  fileDesc("Ch9jbXAvdHlwZXMvdjQvY29udGFjdF9pbmZvLnByb3RvEgxjbXAudHlwZXMudjQizAEKC0NvbnRhY3RJbmZvEjIKCWFkZHJlc3NlcxgBIAMoCzIVLmNtcC50eXBlcy52NC5BZGRyZXNzQgi6SAWSAQIQGRItCgZwaG9uZXMYAiADKAsyEy5jbXAudHlwZXMudjQuUGhvbmVCCLpIBZIBAhAZEi0KBmVtYWlscxgDIAMoCzITLmNtcC50eXBlcy52NC5FbWFpbEIIukgFkgECEBkSKwoFbGlua3MYBCADKAsyEi5jbXAudHlwZXMudjQuTGlua0IIukgFkgECEBliBnByb3RvMw", [file_buf_validate_validate, file_cmp_types_v4_address, file_cmp_types_v4_email, file_cmp_types_v4_link, file_cmp_types_v4_phone]);

/**
 * ContactInfo provides a comprehensive set of contact details for an individual
 * or an organization. It can include multiple physical addresses, phone numbers,
 * email addresses, and web links.
 *
 * This message is designed for general use wherever contact information is needed.
 *
 * ![Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/contact_info.proto.dot.xs.svg)
 *
 * [Open Message Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/contact_info.proto.dot.svg)
 *
 * @generated from message cmp.types.v4.ContactInfo
 */
export type ContactInfo = Message<"cmp.types.v4.ContactInfo"> & {
  /**
   * Optional. A list of physical or postal addresses.
   *
   * Allows for multiple addresses, such as a primary business address,
   * mailing address, or different office locations.
   *
   * @generated from field: repeated cmp.types.v4.Address addresses = 1;
   */
  addresses: Address[];

  /**
   * Optional. A list of phone numbers.
   *
   * Can include various types of phone numbers like mobile, work, fax, etc.
   *
   * @generated from field: repeated cmp.types.v4.Phone phones = 2;
   */
  phones: Phone[];

  /**
   * Optional. A list of email addresses.
   *
   * Allows for multiple email addresses, such as primary contact, support, or
   * departmental emails.
   *
   * @generated from field: repeated cmp.types.v4.Email emails = 3;
   */
  emails: Email[];

  /**
   * Optional. A list of relevant web links or URLs.
   *
   * Can include links to official websites, social media profiles, or other online
   * resources.
   *
   * @generated from field: repeated cmp.types.v4.Link links = 4;
   */
  links: Link[];
};

/**
 * Describes the message cmp.types.v4.ContactInfo.
 * Use `create(ContactInfoSchema)` to create a new message.
 */
export const ContactInfoSchema: GenMessage<ContactInfo> = /*@__PURE__*/
  messageDesc(file_cmp_types_v4_contact_info, 0);